+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file(see http://k
 - map: find value in map, a default value can be provided.
 - sys: retrieve current executable path.
 - sys: retrieve current process id.
+- sys: kill a specific process.
 - regex: add methods support pass std::regex object.
 - regex: add expand method that can use submatch to expand a string.
 - threadpool: user can retrieve the threads count.